+++ src/collective/cover/tiles/list.py
@@ -67,6 +67,7 @@ class IListTile(IPersistentCoverTile):
 
     form.omitted('image')
     form.no_omit(IDefaultConfigureForm, 'image')
+    form.widget(image='collective.cover.tiles.configuration_widgets.namedimage.NamedImageFieldWidget')
     image = NamedBlobImage(
         title=_(u'Image'),
         required=False,